Struct spirit_dipstick::Monitor

pub struct Monitor(_);

An inner node in a metrics tree.

This is just a thin wrapper around the AtomicBucket, hiding some of its functionality. That functionality is used for configuration of the backends and how metrics are collated and filtered, which is handled by this library.

If you want to use the library to just get the configuration and manage it manually, use AtomicBucket directly. This is for the use with Pipelines. See the crate example.

Cloning

Cloning is cheap and creates another „handle“ to the same bucket. Therefore, it can be distributed through the application and another layer of Arc is not needed.

Extraction of AtomicBucket

The bucket inside can be extracted by the Monitor::into_inner method, to access functionality hidden by this wrapper. Note that doing so and messing with the bucket might interfere with configuration from this library (the wrapper is more of a road bump than a full barrier).

Examples

use dipstick::{InputScope, Prefixed};
use spirit_dipstick::Monitor;

let monitor = Monitor::new();
// Plug the monitor.installer() into a pipeline here on startup
let sub_monitor = monitor.add_name("sub");
let timer = sub_monitor.timer("a-timer");
let timer_measurement = timer.start();
let counter = sub_monitor.counter("cnt");
counter.count(1);
timer.stop(timer_measurement);

Implementations

impl Monitor

pub fn new() -> Self

Creates a new root node.

This becomes an independent node. Some backends can be installed into it later on and metrics or other nodes can be created under it.

pub fn into_inner(self) -> AtomicBucket

Extracts the internal AtomicBucket

See the warning above about doing so ‒ the bucket is actually shared by all clones of the same Monitor and therefore you can interfere with the installed backends.

pub fn installer<F>(&self, stats: F) -> MonitorInstaller<F> where
F: Fn(InputKind, MetricName, ScoreType) -> Option<(InputKind, MetricName, MetricValue)> + Send + Sync + 'static,

Creates an installer for installing into this monitor.

This creates an installer which can be used inside a Pipeline to attach backends to it.

The stats is the same kind of filtering and selection function the dipstick uses in the AtomicBucket::stats.